Split KkHakijaUtil to new file
